fix(generator): map int correctly (#37)
* fix(generator): map int correctly

fix #33

* test(basic): fix integer count return type

+++ generator/types/types.go
@@ -35,7 +35,7 @@ var builtin = map[string]string{
 	"ID":       "string",
 	"String":   "string",
 	"Boolean":  "bool",
-	"Int":      "float64", // TODO integer type blocked by https://github.com/prisma/prisma-engine/issues/160
+	"Int":      "int",
 	"Float":    "float64",
 	"DateTime": "DateTime",
 }
